Fixed not displaying all new created FH projects
Before the FH project query was selecting only actual projects, that were
not ended now. This was to keep away very old projects (since 2011...)
But when creating a new FH project by 'copying' the SAP project, the
new FH project was not displayed if the SAP project ending was in the past.

In order to show all SAP projects (even older ones), the query for FH
projects is now selecting also projects closed up to ONE year before
NOW().